+++ lib/animatedNavBar.dart
@@ -1,3 +1,4 @@
+import 'package:animated_navigation_bar/navigation.dart';
 import 'package:flutter/material.dart';
 
 class AnimatedNavBar extends StatefulWidget {
@@ -23,6 +24,7 @@ class AnimatedNavBar extends StatefulWidget {
   /// If the [textStyle] argument is null, the AnimatedNavBar will use the style from the
   /// closest enclosing [DefaultTextStyle].
   final int defaultPage;
+  final ValueChanged<int> selectedPage;
   final bool activeButtonIsWide;
   final Color iconColor;
   final Color inactiveColor;
@@ -36,7 +38,8 @@ class AnimatedNavBar extends StatefulWidget {
       this.padding = 8,
       this.color,
       this.defaultPage = 0,
-        this.activeButtonIsWide = false,
+      this.selectedPage,
+      this.activeButtonIsWide = false,
       this.iconColor,
       this.inactiveColor,
       this.borderRadius,
@@ -60,17 +63,16 @@ class _AnimatedNavBarState extends State<AnimatedNavBar> {
 
   @override
   void initState() {
-    ///
     /// the limit for a [Bottom navigation] destinations is 5 according
     /// to the material.io design guidelines
     /// for more on bottom navigation bars head over to https://material.io/components/bottom-navigation#usage
     ///
     /// if you want to implement more than 5 destinations than a navigation drawer would suffice.
-    if (!(this.widget.pages.length > 0 && this.widget.pages.length < 6))
+    if (!(widget.pages.length > 0 && widget.pages.length < 6))
       throw ("Total count of tabs must not exceed 5 or recede 1");
     size = widget.pages.length;
-    selected = this.widget.defaultPage;
-    Color color = this.widget.color;
+    selected = widget.defaultPage;
+    Color color = widget.color;
     tabButtonHeight = 56;
     shadowColor = Color.fromARGB(120, (color.red * 0.2).toInt(),
         (color.green * 0.2).toInt(), (color.blue * 0.2).toInt());
@@ -86,72 +88,107 @@ class _AnimatedNavBarState extends State<AnimatedNavBar> {
   @override
   Widget build(BuildContext context) {
     return Padding(
-        padding: EdgeInsets.all(this.widget.padding),
-        child: Stack(
-          fit: StackFit.expand,
-          children: [
-            CustomPaint(
+        padding: EdgeInsets.all(widget.padding),
+        child: Stack(fit: StackFit.expand, children: [
+          CustomPaint(
               painter: BackgroundPainter(
+                  (Directionality.of(context) == TextDirection.rtl),
                   selected,
                   size,
-                  this.defaultFlex,
-                  this.widget.color,
+                  defaultFlex,
+                  widget.color,
                   shadowColor,
-                  this.widget.borderRadius.x,
+                  widget.borderRadius.x,
                   tabButtonHeight,
-                  this.widget.shadow),
-            ),
-            Column(
-              mainAxisSize: MainAxisSize.max,
-              children: [
-                // TabView
-                Expanded(
-                  child: ClipPath(
-                    clipper: BackgroundClipper(selected, size,
-                        this.widget.borderRadius.x, this.tabButtonHeight),
-                    child: this.widget.pages[selected].pageContent,
-                  ),
-                ),
-                // TabButtons
-                Row(
-                  children: (() {
-                    List<Widget> _list = [];
-                    for (int i = 0; i < size; i++) {
-                      _list.add(Expanded(
-                          flex: (selected == i) ? this.defaultFlex : 1,
-                          child: Material(
-                            color: Colors.transparent,
-                            child: InkWell(
-                              onTap: () {
-                                setState(() {
-                                  selected = i;
-                                });
-                              },
-                              child: Container(
+                  widget.shadow)),
+          Column(mainAxisSize: MainAxisSize.max, children: [
+            // TabView
+            Expanded(
+                child: ClipPath(
+                    clipper: BackgroundClipper(
+                        (Directionality.of(context) == TextDirection.rtl),
+                        selected,
+                        size,
+                        widget.borderRadius.x,
+                        tabButtonHeight),
+                    child: Stack(
+                        fit: StackFit.expand,
+                        children: (() {
+                          List<Widget> _list = [];
+                          for (int i = 0; i < widget.pages.length; i++) {
+                            _list.add(Offstage(
+                                offstage: selected != i,
+                                child: TabNavigator(
+                                    rootPage: widget.pages[i].pageContent,
+                                    navigatorKey:
+                                        widget.pages[i].navigationKey)));
+                          }
+                          return _list;
+                        }())))),
+            // TabButtons
+            Row(
+                children: (() {
+              List<Widget> _list = [];
+              for (int i = 0; i < size; i++) {
+                _list.add(Expanded(
+                    flex: (selected == i) ? this.defaultFlex : 1,
+                    child: Material(
+                        color: Colors.transparent,
+                        child: InkWell(
+                            onTap: (selected == i)
+                                ? () => setState(() => widget
+                                    .pages[i].navigationKey.currentState
+                                    .popUntil((route) => route.isFirst))
+                                : () => setState(() {
+                                      selected = i;
+                                      widget.selectedPage?.call(selected);
+                                    }),
+                            child: Container(
                                 height: tabButtonHeight,
                                 alignment: Alignment.center,
                                 child: (selected == i)
-                                    ? (this.widget.activeButtonIsWide)
+                                    ? (widget.activeButtonIsWide)
                                         ? Row(
                                             mainAxisAlignment:
                                                 MainAxisAlignment.center,
                                             crossAxisAlignment:
                                                 CrossAxisAlignment.center,
                                             children: [
-                                              Icon(
-                                                this.widget.pages[i].icon,
-                                                color: this.widget.iconColor,
-                                              ),
-                                              SizedBox(
-                                                width: 8,
-                                              ),
-                                              Text(
-                                                this.widget.pages[i].title,
-                                                style: this.widget.textStyle,
-                                                maxLines: 1,
-                                                overflow: TextOverflow.ellipsis,
-                                              )
-                                            ],
+                                                Icon(widget.pages[i].icon,
+                                                    color: widget.iconColor),
+                                                SizedBox(
+                                                  width: 8,
+                                                ),
+                                                Text(widget.pages[i].title,
+                                                    style: widget.textStyle,
+                                                    maxLines: 1,
+                                                    overflow:
+                                                        TextOverflow.ellipsis)
+                                              ])
+                                        : Column(
+                                            crossAxisAlignment:
+                                                CrossAxisAlignment.center,
+                                            mainAxisAlignment:
+                                                MainAxisAlignment.center,
+                                            children: [
+                                                Icon(widget.pages[i].icon,
+                                                    color: widget.iconColor),
+                                                Text(widget.pages[i].title,
+                                                    style: widget.textStyle,
+                                                    maxLines: 1,
+                                                    overflow:
+                                                        TextOverflow.ellipsis)
+                                              ])
+                                    : (widget.activeButtonIsWide)
+                                        ? Icon(
+                                            (widget.pages[i].inactiveIcon !=
+                                                    null)
+                                                ? widget.pages[i].inactiveIcon
+                                                : widget.pages[i].icon,
+                                            color:
+                                                (widget.inactiveColor != null)
+                                                    ? widget.inactiveColor
+                                                    : widget.iconColor,
                                           )
                                         : Column(
                                             crossAxisAlignment:
@@ -160,79 +197,41 @@ class _AnimatedNavBarState extends State<AnimatedNavBar> {
                                                 MainAxisAlignment.center,
                                             children: [
                                                 Icon(
-                                                  this.widget.pages[i].icon,
-                                                  color: this.widget.iconColor,
+                                                  widget.pages[i].icon,
+                                                  color:
+                                                      (widget.inactiveColor !=
+                                                              null)
+                                                          ? widget.inactiveColor
+                                                          : widget.iconColor,
                                                 ),
-                                                Text(
-                                                  this.widget.pages[i].title,
-                                                  style: this.widget.textStyle,
-                                                  maxLines: 1,
-                                                  overflow:
-                                                      TextOverflow.ellipsis,
-                                                )
-                                              ])
-                                    : (this.widget.activeButtonIsWide)
-                                        ? Icon(
-                                            (this
-                                                        .widget
-                                                        .pages[i]
-                                                        .inactiveIcon !=
-                                                    null)
-                                                ? this
-                                                    .widget
-                                                    .pages[i]
-                                                    .inactiveIcon
-                                                : this.widget.pages[i].icon,
-                                            color: (this.widget.inactiveColor !=
-                                                    null)
-                                                ? this.widget.inactiveColor
-                                                : this.widget.iconColor,
-                                          )
-                                        : Column(
-                                            crossAxisAlignment:
-                                                CrossAxisAlignment.center,
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.center,
-                                            children: [
-                                              Icon(
-                                                this.widget.pages[i].icon,
-                                                color: (this
-                                                            .widget
-                                                            .inactiveColor !=
-                                                        null)
-                                                    ? this.widget.inactiveColor
-                                                    : this.widget.iconColor,
-                                              ),
-                                              Text(
-                                                this.widget.pages[i].title,
-                                                style: (this
-                                                            .widget
-                                                            .inactiveColor !=
-                                                        null)
-                                                    ? TextStyle(color: this.widget.inactiveColor)
-                                                    : this.widget.textStyle,
-                                                maxLines: 1,
-                                                overflow: TextOverflow.ellipsis,
-                                              )
-                                            ],
-                                          ),
-                              ),
-                            ),
-                          )));
-                    }
-                    return _list;
-                  }()),
-                )
-              ],
-            ),
-          ],
-        ));
+                                                Text(this.widget.pages[i].title,
+                                                    style: (widget
+                                                                .inactiveColor !=
+                                                            null)
+                                                        ? TextStyle(
+                                                            color: this
+                                                                .widget
+                                                                .inactiveColor)
+                                                        : widget.textStyle,
+                                                    maxLines: 1,
+                                                    overflow:
+                                                        TextOverflow.ellipsis)
+                                              ]))))));
+              }
+              return _list;
+            }()))
+          ])
+        ]));
   }
 }
 
 ///
 /// CustomClipper to clip the Tab view page
 class BackgroundClipper extends CustomClipper<Path> {
+  /// the [isRtl] is to check whether the directionality of the [context]
+  /// is right to left or not, to see from which direction the paint should draw
+  final bool isRtl;
+
   /// [selected] and [tabSize] are to know how to clip the background on the corners
   final int selected;
   final int tabSize;
@@ -243,45 +242,54 @@ class BackgroundClipper extends CustomClipper<Path> {
   /// to get the bottom margin
   final double tabButtonHeight;
 
-  BackgroundClipper(
-      this.selected, this.tabSize, this.radius, this.tabButtonHeight);
+  BackgroundClipper(this.isRtl, this.selected, this.tabSize, this.radius,
+      this.tabButtonHeight);
 
   @override
   Path getClip(Size size) {
     Path path = Path();
 
-    /// starting point : right side
-    path.moveTo(size.width, radius);
+    /// to control which direction the clipper should start clipping
+    /// depending on the directionality ot the [Locale]
+    double extremeRight = size.width * (isRtl ? 0 : 1),
+        extremeLeft = size.width * (isRtl ? 1 : 0),
+        directionalitySign = (isRtl ? 1 : -1);
 
-    /// top right corner
-    path.quadraticBezierTo(size.width, 0, size.width - radius, 0);
+    /// starting point : right side, left for rtl
+    path.moveTo(extremeRight, radius);
+
+    /// top right corner, top left for rtl
+    path.quadraticBezierTo(
+        extremeRight, 0, extremeRight + directionalitySign * radius, 0);
 
     /// top line
-    path.lineTo(radius, 0);
+    path.lineTo(extremeLeft - directionalitySign * radius, 0);
 
-    /// top left corner
-    path.quadraticBezierTo(0, 0, 0, radius);
+    /// top left corner, rop right for rtl
+    path.quadraticBezierTo(extremeLeft, 0, extremeLeft, radius);
 
-    /// bottom left corner
+    /// bottom left corner, bottom right in case of [rtl]
     /// if selected item is the first one
     if (selected == 0) {
-      path.lineTo(0, size.height);
-      path.lineTo(size.width - radius, size.height);
+      path.lineTo(extremeLeft, size.height);
+      path.lineTo(extremeRight + directionalitySign * radius, size.height);
       path.quadraticBezierTo(
-          size.width, size.height, size.width, size.height - radius);
+          extremeRight, size.height, extremeRight, size.height - radius);
     } else if (selected == tabSize - 1) {
-      path.lineTo(0, size.height - radius);
-      path.quadraticBezierTo(0, size.height, radius, size.height);
-      path.lineTo(size.width, size.height);
+      path.lineTo(extremeLeft, size.height - radius);
+      path.quadraticBezierTo(extremeLeft, size.height,
+          extremeLeft - directionalitySign * radius, size.height);
+      path.lineTo(extremeRight, size.height);
     } else {
-      path.lineTo(0, size.height - radius);
-      path.quadraticBezierTo(0, size.height, radius, size.height);
-      path.lineTo(size.width - radius, size.height);
+      path.lineTo(extremeLeft, size.height - radius);
+      path.quadraticBezierTo(extremeLeft, size.height,
+          extremeLeft - directionalitySign * radius, size.height);
+      path.lineTo(extremeRight + directionalitySign * radius, size.height);
       path.quadraticBezierTo(
-          size.width, size.height, size.width, size.height - radius);
+          extremeRight, size.height, extremeRight, size.height - radius);
     }
 
-    path.lineTo(size.width, radius);
+    path.lineTo(extremeRight, radius);
     return path;
   }
 
@@ -294,17 +302,32 @@ class BackgroundClipper extends CustomClipper<Path> {
 ///
 /// CustomPainter to paint the Tab view background
 class BackgroundPainter extends CustomPainter {
+  /// the [isRtl] is to check whether the directionality of the [context]
+  /// is right to left or not, to see from which direction the paint should draw
+  final bool isRtl;
+
+  /// [selected] and [tabSize] are to know how to clip the background on the corners
   final int selected;
   final int tabSize;
   final int defaultFlex;
   final Color color;
   final Color shadowColor;
+
+  /// [radius] is the border radius value
   final double radius;
   final double tabButtonHeight;
   final bool shadow;
 
-  BackgroundPainter(this.selected, this.tabSize, this.defaultFlex, this.color,
-      this.shadowColor, this.radius, this.tabButtonHeight, this.shadow);
+  BackgroundPainter(
+      this.isRtl,
+      this.selected,
+      this.tabSize,
+      this.defaultFlex,
+      this.color,
+      this.shadowColor,
+      this.radius,
+      this.tabButtonHeight,
+      this.shadow);
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -316,24 +339,34 @@ class BackgroundPainter extends CustomPainter {
 
     Path path = Path();
 
+    /// to control which direction the painter should start painting
+    /// depending on the directionality ot the [Locale]
+    double extremeRight = size.width * (isRtl ? 0 : 1),
+        extremeLeft = size.width * (isRtl ? 1 : 0),
+        directionalitySign = (isRtl ? 1 : -1);
+
     /// starting point : right side
-    path.moveTo(size.width, radius);
+    path.moveTo(extremeRight, radius);
 
     /// top right corner
-    path.quadraticBezierTo(size.width, 0, size.width - radius, 0);
+    path.quadraticBezierTo(
+        extremeRight, 0, extremeRight + directionalitySign * radius, 0);
 
     /// top line
-    path.lineTo(radius, 0);
+    path.lineTo(extremeLeft - directionalitySign * radius, 0);
 
     /// top left corner
-    path.quadraticBezierTo(0, 0, 0, radius);
+    path.quadraticBezierTo(extremeLeft, 0, extremeLeft, radius);
 
     /// bottom left corner
     if (selected == 0)
-      path.lineTo(0, size.height - tabButtonHeight);
+      path.lineTo(extremeLeft, size.height - tabButtonHeight);
     else {
-      path.lineTo(0, size.height - tabButtonHeight - radius);
-      path.quadraticBezierTo(0, size.height - tabButtonHeight, radius,
+      path.lineTo(extremeLeft, size.height - tabButtonHeight - radius);
+      path.quadraticBezierTo(
+          extremeLeft,
+          size.height - tabButtonHeight,
+          extremeLeft - directionalitySign * radius,
           size.height - tabButtonHeight);
     }
 
@@ -341,44 +374,62 @@ class BackgroundPainter extends CustomPainter {
     for (int i = 0; i < tabSize; i++) {
       /// if painter is before selected item
       if (i == selected - 1) {
-        path.lineTo(flexSize * (i + 1) - radius, size.height - tabButtonHeight);
+        path.lineTo(
+            extremeLeft - directionalitySign * (flexSize * (i + 1) - radius),
+            size.height - tabButtonHeight);
         path.quadraticBezierTo(
-            flexSize * (i + 1),
+            extremeLeft - directionalitySign * (flexSize * (i + 1)),
             size.height - tabButtonHeight,
-            flexSize * (i + 1),
+            extremeLeft - directionalitySign * (flexSize * (i + 1)),
             size.height - tabButtonHeight + radius);
       }
 
       /// if painter is on selected item
       if (i == selected) {
-        path.lineTo(flexSize * i, size.height - radius);
+        path.lineTo(extremeLeft - directionalitySign * (flexSize * i),
+            size.height - radius);
         path.quadraticBezierTo(
-            flexSize * i, size.height, flexSize * i + radius, size.height);
-        path.lineTo(flexSize * i + activeTabWidth - radius, size.height);
-        path.quadraticBezierTo(flexSize * i + activeTabWidth, size.height,
-            flexSize * i + activeTabWidth, size.height - radius);
+            extremeLeft - directionalitySign * (flexSize * i),
+            size.height,
+            extremeLeft - directionalitySign * (flexSize * i + radius),
+            size.height);
+        path.lineTo(
+            extremeLeft -
+                directionalitySign * (flexSize * i + activeTabWidth - radius),
+            size.height);
+        path.quadraticBezierTo(
+            extremeLeft - directionalitySign * (flexSize * i + activeTabWidth),
+            size.height,
+            extremeLeft - directionalitySign * (flexSize * i + activeTabWidth),
+            size.height - radius);
       }
 
       /// if painter is after selected item
       if (i == selected + 1) {
-        path.lineTo(flexSize * (i - 1) + activeTabWidth,
+        path.lineTo(
+            extremeLeft -
+                directionalitySign * (flexSize * (i - 1) + activeTabWidth),
             size.height - tabButtonHeight + radius);
         path.quadraticBezierTo(
-            flexSize * (i - 1) + activeTabWidth,
+            extremeLeft -
+                directionalitySign * (flexSize * (i - 1) + activeTabWidth),
             size.height - tabButtonHeight,
-            flexSize * (i - 1) + activeTabWidth + radius,
+            extremeLeft -
+                directionalitySign *
+                    (flexSize * (i - 1) + activeTabWidth + radius),
             size.height - tabButtonHeight);
       }
     }
 
     /// bottom right corner
     if (selected == tabSize - 1)
-      path.lineTo(size.width, radius);
+      path.lineTo(extremeRight, radius);
     else {
-      path.lineTo(size.width - radius, size.height - tabButtonHeight);
-      path.quadraticBezierTo(size.width, size.height - tabButtonHeight,
-          size.width, size.height - tabButtonHeight - radius);
-      path.lineTo(size.width, radius);
+      path.lineTo(extremeRight + directionalitySign * radius,
+          size.height - tabButtonHeight);
+      path.quadraticBezierTo(extremeRight, size.height - tabButtonHeight,
+          extremeRight, size.height - tabButtonHeight - radius);
+      path.lineTo(extremeRight, radius);
     }
 
     /// drawing the canvas
@@ -402,7 +453,13 @@ class AnimatedNavBarPage {
   final Widget pageContent;
   final IconData icon;
   final IconData inactiveIcon;
+  final GlobalKey<NavigatorState> navigationKey;
 
   const AnimatedNavBarPage(
-      {Key key, this.title, this.pageContent, this.icon, this.inactiveIcon});
+      {Key key,
+      @required this.title,
+      @required this.pageContent,
+      @required this.navigationKey,
+      this.icon,
+      this.inactiveIcon});
 }