+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2021-11-03
+- Added Right to Left Directionality and Locale support
+- Added PopUntil Root when pressing on an active tab button
+
 ## [0.0.3+1] - 2021-09-22
 - Fixed a bug